requirements.txt
numpy==2.3.2
pandas==2.3.1
requests==2.32.4
scholarly==1.7.11
SQLAlchemy==2.0.43
streamlit==1.48.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
